refactor: Split class and improve streams usage
